move execution history helpers to new file
