+++ packages/common/tests/test-sfnStep.js
@@ -6,69 +6,8 @@ const rewire = require('rewire');
 const SfnStepModule = rewire('../sfnStep');
 
 const {
-  SfnStep,
-  getStepExitedEvent,
-  getTaskExitedEventOutput
+  SfnStep
 } = SfnStepModule;
-
-test('getStepExitedEvent returns undefined if task exited event cannot be found', (t) => {
-  const events = [
-    {
-      type: 'LambdafunctionFailed',
-      id: 1,
-      lambdaFunctionFailedEventDetails: {}
-    },
-    {
-      type: 'TaskStateExited',
-      id: 2,
-      previousEventId: 1,
-      stateExitedEventDetails: {
-        output: {}
-      }
-    }
-  ];
-
-  t.is(getStepExitedEvent(events, { id: 3 }), undefined);
-});
-
-test('getStepExitedEvent returns correct task exited event', (t) => {
-  const events = [
-    {
-      type: 'LambdafunctionFailed',
-      id: 1,
-      lambdaFunctionFailedEventDetails: {}
-    },
-    {
-      type: 'TaskStateExited',
-      id: 2,
-      previousEventId: 1,
-      stateExitedEventDetails: {
-        output: {}
-      }
-    }
-  ];
-
-  const taskExitedEvent = getStepExitedEvent(
-    events,
-    { id: 1 }
-  );
-  t.truthy(taskExitedEvent);
-  t.is(taskExitedEvent.type, 'TaskStateExited');
-  t.truthy(taskExitedEvent.stateExitedEventDetails);
-});
-
-test('getTaskExitedEventOutput returns correct output', (t) => {
-  const payload = {
-    foo: 'bar'
-  };
-  const event = {
-    stateExitedEventDetails: {
-      output: JSON.stringify(payload)
-    }
-  };
-  const output = getTaskExitedEventOutput(event);
-  t.deepEqual(JSON.parse(output), payload);
-});
 
 test('SfnStep.parseStepMessage parses message correctly', async (t) => {
   const event = {
